Fixes for client feedback on the Collection Form.
Changed language code autocomplete to search by title rather than by
code. Now shows both the title and the code in the search results.

Now setting the a field linked with a language like indigenous title,
will fail to submit if the language is not also given. Though setting
the language without giving a title yields nothing.

Fixes for empty fields on update:

* Title (English)
* Title (Spanish)
* Title (Indigenous language)
* Title (Indigenous language) Language
* Collection description (English)
* Collection description (Spanish)
* Collection description (Indigenous) Language